+++ src/main/java/com/pollra/pudding/web/business/comment/entity/Comment.java
@@ -1,7 +1,7 @@
-package com.pollra.pudding.web.comment.entity;
+package com.pollra.pudding.web.business.comment.entity;
 
 import com.pollra.pudding.common.base.entity.Sign;
-import com.pollra.pudding.web.comment.enumerated.CommentStatus;
+import com.pollra.pudding.web.business.comment.enumerated.CommentStatus;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
