+++ src/main/java/com/pollra/pudding/web/business/tag/entity/TagPool.java
@@ -1,4 +1,4 @@
-package com.pollra.pudding.web.tag.entity;
+package com.pollra.pudding.web.business.tag.entity;
 
 import com.pollra.pudding.common.base.entity.Sign;
 
@@ -7,12 +7,10 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
 @Entity
-public class Tag extends Sign {
+public class TagPool extends Sign {
     @Id @GeneratedValue
     private Long id;
-    private Long poolId;
     private Long aclId;
-    private Long postId;
     private String name;
     private String description;
 }
